all review queries and mutations complete

+++ server/schemas/resolvers.js
@@ -26,10 +26,13 @@ const resolvers = {
             return await User.find().populate('review').sort({createdAt: -1});
         },
         getUserById: async (parent,{_id}) => {
-            return await User.findOne({_id:_id});
+            return await User.findOne({_id:_id}).populate('review');
         },
         getReviews: async  (parent,args) => {
             return await Review.find().populate('author').sort({createdAt: -1});
+        },
+        getReviewById: async (parent,{_id}) => {
+            return await Review.findOne({_id:_id}).populate('author');
         },
     },
     Mutation: {
@@ -67,13 +70,28 @@ const resolvers = {
             throw new AuthenticationError('no permissions');
         },
         deleteUser: async (parent,{_id},context) => {
-          // if(context.user._id === _id) {
+           if(context.user._id === _id) {
             const user = await User.findById(_id);
-            console.log(user);
+            if(user.hasLeftReview === true){
+            try{
+                await Review.findByIdAndDelete(user.review._id);
                 const deletedUser = await User.findOneAndDelete({_id:_id});
+                console.log('deleting user and review block ran');
                 return deletedUser;
-           // }
-           // throw new AuthenticationError('it must actually be you account to delete!');
+            }catch(e) {
+                console.log('delete review and user block',e);
+             }
+            }
+               try{
+                const deletedUser = await User.findOneAndDelete({_id:_id});
+                console.log('only deleting user block ran');
+                return deletedUser;
+               }catch(e) {
+                console.log('only delete user block',e);
+               }
+        
+            }
+           throw new AuthenticationError('it must actually be you account to delete!');
         },
         addReview: async (parent,{reviewText,rating},context) => {
             const currentUser = await User.findOne({_id:context.user._id});
@@ -97,9 +115,46 @@ const resolvers = {
             }
             throw new AuthenticationError('looks like you have already left a review!');
         },
-        deleteReview: async (parent,{_id}) => {
+        deleteReview: async (parent,{_id},context) => {
+            /*find review thats fed in to mutation*/
+            const review = await Review.findById(_id);
+            /*get author's id from the found review*/
+            const author = review.author.toHexString();
+            /*here we compare author id to the current users id if they match delete review if not throw error */
+            if(author === context.user._id ) {
             const deletedReview = await Review.findOneAndDelete({_id:_id});
+            try{await User.findByIdAndUpdate({_id:context.user._id},{hasLeftReview:false},{new:true});}
+            catch(e){
+                console.log('update user fail review delete',e);
+            }
             return deletedReview;
+            }
+            throw new AuthenticationError('you must have made the review to delete it!')
+        },
+        updateReview: async (parent,{_id,reviewText,rating},context) => {
+              /*find review thats fed in to mutation*/
+              const review = await Review.findById(_id);
+              /*get author's id from the found review*/
+            const author = review.author.toHexString();
+            const currentUser = context.user._id;
+            console.log(author);
+            console.log(currentUser);
+            if(author === currentUser) {
+                try {
+                const updatedReview = await Review.findByIdAndUpdate(
+                    {_id: _id},
+                    {
+                        reviewText: reviewText,
+                        rating: rating
+                    },
+                    {new:true} 
+                );
+                return updatedReview;
+                }catch(e) {
+                    console.log(e);
+                }
+            }
+          throw new AuthenticationError('to edit a review you must have created it and be logged in!');
         },
     }
 };
